+++ kafka-junit5/src/main/java/com/salesforce/kafka/test/junit5/SharedKafkaTestResource.java
@@ -28,20 +28,57 @@ package com.salesforce.kafka.test.junit5;
 import com.salesforce.kafka.test.KafkaTestServer;
 import com.salesforce.kafka.test.KafkaTestUtils;
 import org.apache.curator.test.TestingServer;
+import org.junit.jupiter.api.extension.AfterAllCallback;
+import org.junit.jupiter.api.extension.BeforeAllCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.Properties;
 
 /**
- * Shared Kafka Test Resource instance.  Contains references to internal Kafka and Zookeeper server instances.
+ * Creates and stands up an internal test kafka server to be shared across test cases within the same test class.
+ *
+ * Example within your Test class.
+ *
+ *   &#064;RegisterExtension
+ *   public static final SharedKafkaTestResource sharedKafkaTestResource = new SharedKafkaTestResource();
+ *
+ * Within your test case method:
+ *   sharedKafkaTestResource.getKafkaTestServer()...
  */
-public class SharedKafkaTestResource {
+public class SharedKafkaTestResource implements BeforeAllCallback, AfterAllCallback {
+    private static final Logger logger = LoggerFactory.getLogger(SharedKafkaTestResource.class);
+
     /**
      * Our internal Kafka Test Server instance.
      */
-    private final KafkaTestServer kafkaTestServer = new KafkaTestServer();
+    private KafkaTestServer kafkaTestServer = null;
 
     /**
      * Cached instance of KafkaTestUtils.
      */
     private KafkaTestUtils kafkaTestUtils = null;
+
+    /**
+     * Additional broker properties.
+     */
+    private final Properties brokerProperties = new Properties();
+
+    /**
+     * Default constructor.
+     */
+    public SharedKafkaTestResource() {
+        this(new Properties());
+    }
+
+    /**
+     * Constructor allowing passing additional broker properties.
+     * @param brokerProperties properties for Kafka broker.
+     */
+    public SharedKafkaTestResource(final Properties brokerProperties) {
+        this.brokerProperties.putAll(brokerProperties);
+    }
 
     /**
      * @return Shared Kafka Test server instance.
@@ -79,5 +116,68 @@ public class SharedKafkaTestResource {
      */
     public String getKafkaConnectString() {
         return getKafkaTestServer().getKafkaConnectString();
+    }
+
+    /**
+     * Helper to allow overriding Kafka broker properties.  Can only be called prior to the service
+     * being started.
+     * @param name Kafka broker configuration property name.
+     * @param value Value to set for the configuration property.
+     * @return SharedKafkaTestResource instance for method chaining.
+     * @throws IllegalArgumentException if name argument is null.
+     * @throws IllegalStateException if method called after service has started.
+     */
+    public SharedKafkaTestResource withBrokerProperty(final String name, final String value) {
+        // Validate input.
+        if (name == null) {
+            throw new IllegalArgumentException("Cannot pass null name argument");
+        }
+
+        // Validate state.
+        if (kafkaTestServer != null) {
+            throw new IllegalStateException("Cannot add properties after service has started");
+        }
+
+        // Add or set property.
+        if (value == null) {
+            brokerProperties.remove(name);
+        } else {
+            brokerProperties.put(name, value);
+        }
+        return this;
+    }
+
+    /**
+     * Here we stand up an internal test kafka and zookeeper service.
+     * Once for all tests that use this shared resource.
+     */
+    @Override
+    public void beforeAll(ExtensionContext context) throws Exception {
+        logger.info("Starting kafka test server");
+        if (kafkaTestServer != null) {
+            throw new IllegalStateException("Unknown State!  Kafka Test Server already exists!");
+        }
+        // Setup kafka test server
+        kafkaTestServer = new KafkaTestServer(brokerProperties);
+        kafkaTestServer.start();
+    }
+
+    /**
+     * Here we shut down the internal test kafka and zookeeper services.
+     */
+    @Override
+    public void afterAll(ExtensionContext context) throws Exception {
+        logger.info("Shutting down kafka test server");
+
+        // Close out kafka test server if needed
+        if (kafkaTestServer == null) {
+            return;
+        }
+        try {
+            kafkaTestServer.shutdown();
+        } catch (final Exception e) {
+            throw new RuntimeException(e);
+        }
+        kafkaTestServer = null;
     }
 }