+++ kafka-junit5/src/test/java/com/salesforce/kafka/test/junit5/ZookeeperResourseExtensionTest.java
@@ -1,0 +1,82 @@
+/**
+ * Copyright (c) 2017-2018, Salesforce.com, Inc.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
+ * following conditions are met:
+ *
+ * * Redistributions of source code must retain the above copyright notice, this list of conditions and the following
+ *   disclaimer.
+ *
+ * * Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following
+ *   disclaimer in the documentation and/or other materials provided with the distribution.
+ *
+ * * Neither the name of Salesforce.com nor the names of its contributors may be used to endorse or promote products
+ *   derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
+ * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+ * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.salesforce.kafka.test.junit5;
+
+import org.apache.curator.test.TestingServer;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+/**
+ * This serves as an example of how to use this library using @ExtendWith annotation.
+ *
+ * @deprecated Please use SharedZookeeperTestResourceTest as a reference for using this library.
+ */
+@Deprecated
+@ExtendWith(ZookeeperResourceExtension.class)
+public class ZookeeperResourseExtensionTest {
+    /**
+     * We have a single embedded zookeeper server that gets started when this test class is initialized.
+     *
+     * It's automatically started before any methods are run via the @ExtendWith annotation.
+     * It's automatically stopped after all of the tests are completed via the @ExtendWith annotation.
+     * This instance is passed to this class's constructor via the @ExtendWith annotation.
+     */
+    private final SharedZookeeperTestResource sharedZookeeperTestResource;
+
+    /**
+     * Constructor where KafkaResourceExtension provides the sharedKafkaTestResource object.
+     * @param sharedZookeeperTestResource Provided by ZookeeperResourceExtension.
+     */
+    public ZookeeperResourseExtensionTest(final SharedZookeeperTestResource sharedZookeeperTestResource) {
+        this.sharedZookeeperTestResource = sharedZookeeperTestResource;
+    }
+
+    /**
+     * Validates that we receive a sane looking ZK connection string.
+     */
+    @Test
+    void testGetZookeeperConnectString() {
+        final String actualConnectStr = sharedZookeeperTestResource.getZookeeperConnectString();
+
+        // Validate
+        assertNotNull(actualConnectStr, "Should have non-null connect string");
+        assertTrue(actualConnectStr.startsWith("127.0.0.1:"), "Should start with 127.0.0.1");
+    }
+
+    /**
+     * Validates that we receive a sane looking ZK connection string.
+     */
+    @Test
+    void testZookeeperServer() {
+        final TestingServer zkTestServer = sharedZookeeperTestResource.getZookeeperTestServer();
+
+        // Validate
+        assertNotNull(zkTestServer, "Should have non-null instance");
+    }
+}